Insert special symbol when code is broken by transformation.
Plus: extract helpers for code transforation.
